# Conflicts:
#	app/admin_room.py
#	app/templates/admin_room.html
#	app/templates/main_tab.html
#	app/test.db

+++ app/user_delete.py
@@ -2,14 +2,17 @@ from flask import render_template, request, redirect, url_for, Blueprint, Flask,
 
 from models.User import User
 from models.database import db
+from login import login_required
 user_delete_bp = Blueprint('user_delete', __name__, url_prefix='/user_delete')
 
 @user_delete_bp.route('/<user_email>')
 def user_delete(user_email):
-    if user_email != session['user']:
-        user = db.session.query(User).filter_by(email=user_email).first()
-        db.session.delete(user)
-        db.session.commit()
+    if login_required():
+        if user_email != session['user']:
+            user = db.session.query(User).filter_by(email=user_email).first()
+            db.session.delete(user)
+            db.session.commit()
+            
+        return redirect(url_for('admin_main.admin_main'))
+    return redirect(url_for('login.login'))
 
-
-    return redirect(url_for('admin_main.admin_main'))
